requirements.txt
discord.py
nest-asyncio
